requirements.txt
streamlit
langchain
langchain-openai
langchain-community
pinecone-client
pypdf
python-dotenv
